refactor: move function
Since the `tryGetTargetEditorVersionFor` function now acts on package-manifests it makes sense to move it into the corresponding file.
