+++ test/domain/package-manifest.test.ts
@@ -1,4 +1,15 @@
-import { dependenciesOf } from "../../src/domain/package-manifest";
+import {
+  dependenciesOf,
+  tryGetTargetEditorVersionFor,
+} from "../../src/domain/package-manifest";
+import fc from "fast-check";
+import { arbDomainName } from "./domain-name.arb";
+import { makeEditorVersion } from "../../src/domain/editor-version";
+import {
+  InvalidTargetEditorError,
+  UnityPackumentVersion,
+} from "../../src/domain/packument";
+import { makeSemanticVersion } from "../../src/domain/semantic-version";
 
 describe("package manifest", () => {
   describe("get dependency list", () => {
@@ -26,6 +37,78 @@ describe("package manifest", () => {
       const dependencies = dependenciesOf(packageManifest);
 
       expect(dependencies).toEqual([]);
+    });
+  });
+
+  describe("determine editor-version", () => {
+    it("should get version without release", () => {
+      fc.assert(
+        fc.property(arbDomainName, (packumentName) => {
+          const expected = makeEditorVersion(2020, 3);
+          const packumentVersion: UnityPackumentVersion = {
+            name: packumentName,
+            version: makeSemanticVersion("1.0.0"),
+            unity: "2020.3",
+          };
+
+          const result = tryGetTargetEditorVersionFor(packumentVersion);
+
+          expect(result).toBeOk((actual) => expect(actual).toEqual(expected));
+        })
+      );
+    });
+
+    it("should get version with release", () => {
+      fc.assert(
+        fc.property(arbDomainName, (packumentName) => {
+          const expected = makeEditorVersion(2020, 3, 1);
+          const packumentVersion: UnityPackumentVersion = {
+            name: packumentName,
+            version: makeSemanticVersion("1.0.0"),
+            unity: "2020.3",
+            unityRelease: "1",
+          };
+
+          const result = tryGetTargetEditorVersionFor(packumentVersion);
+
+          expect(result).toBeOk((actual) => expect(actual).toEqual(expected));
+        })
+      );
+    });
+
+    it("should be null for missing major.minor", () => {
+      fc.assert(
+        fc.property(arbDomainName, (packumentName) => {
+          const packumentVersion: UnityPackumentVersion = {
+            name: packumentName,
+            version: makeSemanticVersion("1.0.0"),
+          };
+
+          const result = tryGetTargetEditorVersionFor(packumentVersion);
+
+          expect(result).toBeOk((actual) => expect(actual).toBeNull());
+        })
+      );
+    });
+
+    it("should fail for bad version", () => {
+      fc.assert(
+        fc.property(arbDomainName, (packumentName) => {
+          const packumentVersion: UnityPackumentVersion = {
+            name: packumentName,
+            version: makeSemanticVersion("1.0.0"),
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-ignore
+            unity: "bad version",
+          };
+
+          const result = tryGetTargetEditorVersionFor(packumentVersion);
+
+          expect(result).toBeError((actual) =>
+            expect(actual).toBeInstanceOf(InvalidTargetEditorError)
+          );
+        })
+      );
     });
   });
 });
